Rename map_lifting.rs to be consistent with its contents
