+++ src/transform/src/lib.rs
@@ -39,7 +39,7 @@ pub mod demand;
 pub mod fusion;
 pub mod inline_let;
 pub mod join_implementation;
-pub mod map_lifting;
+pub mod literal_lifting;
 pub mod monotonic;
 pub mod nonnull_requirements;
 pub mod nonnullable;
@@ -359,7 +359,7 @@ impl Optimizer {
                     // Converts `Cross Join {Constant(Literal) + Input}` to
                     // `Map {Cross Join (Input, Constant()), Literal}`.
                     // Join fusion will clean this up to `Map{Input, Literal}`
-                    Box::new(crate::map_lifting::LiteralLifting::default()),
+                    Box::new(crate::literal_lifting::LiteralLifting::default()),
                     // Identifies common relation subexpressions.
                     // Must be followed by let inlining, to keep under control.
                     Box::new(crate::cse::relation_cse::RelationCSE),
@@ -394,7 +394,7 @@ impl Optimizer {
                     Box::new(crate::column_knowledge::ColumnKnowledge::default()),
                     Box::new(crate::reduction::FoldConstants { limit: Some(10000) }),
                     Box::new(crate::demand::Demand::default()),
-                    Box::new(crate::map_lifting::LiteralLifting::default()),
+                    Box::new(crate::literal_lifting::LiteralLifting::default()),
                 ],
             }),
             Box::new(crate::canonicalize_mfp::CanonicalizeMfp),
