+++ src/transform/tests/test_runner.rs
@@ -239,7 +239,7 @@ mod tests {
             "FlatMapToMap" => Ok(Box::new(mz_transform::fusion::flatmap_to_map::FlatMapToMap)),
             "JoinFusion" => Ok(Box::new(mz_transform::fusion::join::Join)),
             "LiteralLifting" => Ok(Box::new(
-                mz_transform::map_lifting::LiteralLifting::default(),
+                mz_transform::literal_lifting::LiteralLifting::default(),
             )),
             "NonNullRequirements" => Ok(Box::new(
                 mz_transform::nonnull_requirements::NonNullRequirements::default(),
